pyproject.toml
[tool.poetry]
name = "econoapp"
version = "0.1.0"
description = "App de finanças pessoais"
authors = ["Oseias Nascimento <oseiasal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
